Extracted constants from Filters.jsx and moved them to filters.js
